Fix remaining issues with social image capture
